requirements.txt
httpx
paho-mqtt
trio
isort
black
